requirements.txt
python-telegram-bot==22.5
Pillow==11.3.0
fastapi==0.117.1